Reorder the datatype definition and use a storage class coprresponding to the kind of data it's supposed to hold. Change the stack usage to make things faster:  - the first of the stack is always used by the count argument  - the second is initialized to the first position in the elements list.  - when the ddt engine advance the last stack position hold the informations about the last processed element (the one from where we have to start the next time).
  Modify the pack/unpack relative functions to reflect the changes in the stack usage. In same time I decrease the amount of code (yup yup) and reorder it a little bit. I get some speedup on my laptop ...

This commit was SVN r3221.

+++ src/datatype/dt_sndrcv.c
@@ -55,11 +55,11 @@ int ompi_ddt_sndrcv(void *sbuf, int scount, MPI_Datatype sdtype, void *rbuf,
 				      scount, NULL, 0, NULL );
         err = ompi_convertor_get_packed_size(local_convertor, &size);
         OBJ_RELEASE(local_convertor);
-        if (OMPI_SUCCESS != err) {
+        if( OMPI_SUCCESS != err ) {
             return err;
         }
 
-	if (size <= rcount) {
+	if( (int)size <= rcount ) {
 	    err = MPI_Pack(sbuf, scount, sdtype,
                            rbuf, rcount, &position, MPI_COMM_WORLD);
 	} else {
@@ -75,11 +75,11 @@ int ompi_ddt_sndrcv(void *sbuf, int scount, MPI_Datatype sdtype, void *rbuf,
                                      rcount, NULL, 0, NULL );
         err = ompi_convertor_get_packed_size(local_convertor, &size);
         OBJ_RELEASE(local_convertor);
-        if (OMPI_SUCCESS != err) {
+        if( OMPI_SUCCESS != err ) {
             return err;
         }
 
-	if (scount <= size) {
+	if( scount <= (int)size ) {
 	    err = MPI_Unpack(sbuf, scount, &position,
                              rbuf, rcount, rdtype, 
                              MPI_COMM_WORLD);
